Rewrote trotStack, factored our memory functions, started trotListSecondary.c, fixed Makefile.
Rewrote trotStack because it needed additional functionality for the upcoming trotListSecondary functions.

Factored the memory functions out into their own file. In the future we may add additional memory related functions.

Created trotListSecondary.c, which is the next phase of development.

Fixed Makefile in trotTest. It was not depending on the trot library.