Changing our "free" function so that it doesn't allocate memory to free memory.
We previously used a stack to remember which lists we visited so we didn't get
into a circular loop. That required allocating memory to grow the stack. Now we
keep some variables in our list to accomplish the same thing.

Also fixed our "debug2" make target and added "-g" to our debug targets so
valgrind prints out line numbers.