Comments upgraded
I previously thought comments would be handled as meta-data, but that turned
out to be a bad idea. If we did that, then the contents of comments would
have to be comprised of valid tokens.
So, for example, this wouldn't work:
(comment this comment has a stray " double quote)

However, the new comments are handled in the tokenizer, which removes this
problem.
A single hash '#' creates a comment to the end of the line.

 # a single line comment to end of line
 # another single line comment

Hashes with '(' or ')' create grouped comments that can be nested.

 #( this is a simple
    multi-line comment #)

 #( another simple
    multi-line comment
 #)

 #( this is a comment
    this is still a comment
 #( more comments #( even more here #) and here
 #) still a comment here because they can be nested
 last of comment
 #)