Implements getTag setTag functions
Also had to make sure our enums started at 1 instead of 0.
It turns out the compiler can decide to make enum values signed or unsigned.
If the enum is unsigned, and we want to make sure an enum value is good, we might think to do this:
if ( someValue < MIN_ENUM_VALUE )
But if the enum type is unsigned, and MIN_ENUM_VALUE is 0, the compiler will give a warning that
comparing an unsigned value to be less than 0 is always false.
If we start enum values at 1 instead, we can still check enum values at both high and low boundaries.

Ran all tests, we're still at 100% branch coverage.