Improves handling include metadata
(include) must be the last token in a list.
This, coupled with the check that metadata must come first in a list
guarantees that lists that are includes won't contain any other data.

lists that are (include) must be inside a bracket. The file it
includes may have braces as it's top-most token, and so the list may
eventually become tagged as code (braces), but includes must always
be a "generic" bracket list.

Quality:
Compiles with no warnings or errors.
All unit tests pass.
100% branch coverage.