Allows names to go "through" includes.
Let's say we have two files.

FILE A:
[(name A)
	[(name B)(include "fileB")
	A.B.childList
	A.B.childList.enum1
]

FILE B:
[
	[(name childList)(enum (enum1 1))
		1 2 3
	]
]

Notice that in file A we have this name "A.B.childList".
The "A.B" part references the include, and therefore file B.
the next "childList" part references a child in file B.

A dotted name should be able to go "through" an include into
the included file. I never tested or implemented this.

This change implements and tests this.

Quality:
Compiles with no warnings or errors.
All unit tests pass.
100% code coverage.