Massive change in PLAN with compiler.
Instead of creating the compiler in C, it will be much easier and faster to
create it in Trot itself. So I'm going to create a very basic compiler in C,
which is essentially the most bare-bones encoder and decoder possible, and then
later write the real full-functioning compiler in Trot. All the work I've done
with the compiler so far in C isn't wasted, as I can still use the structure as
a reference when I'm creating the Trot version.

All tests pass, but still need to get 100% branch coverage for the modified
trotDecoding.c and trotEncoding.c.